refactorizado completo
