+++ src/tests/Home/WsHomeService.test.js
@@ -1,90 +1,91 @@
-import { renderHook, act } from '@testing-library/react';
-import WsHomeService from '../../services/WsHomeService';
-import { afterAll } from 'vitest';
+import { renderHook, act } from "@testing-library/react";
+import WsHomeService from "../../services/WS/WsHomeService";
+import { afterAll } from "vitest";
 
+describe("WsHomeService", () => {
+  const url = "ws://localhost:1234";
+  let mockWebSocket;
+  const consoleLogSpy = vi.spyOn(console, "log").mockImplementation(() => {});
+  const consoleErrorSpy = vi
+    .spyOn(console, "error")
+    .mockImplementation(() => {});
 
-describe('WsHomeService', () => {
-    const url = 'ws://localhost:1234';
-    let mockWebSocket;
-    const consoleLogSpy = vi.spyOn(console, 'log').mockImplementation(() => {});
-    const consoleErrorSpy = vi.spyOn(console, 'error').mockImplementation(() => {});
+  beforeEach(() => {
+    mockWebSocket = {
+      send: vi.fn(),
+      close: vi.fn(),
+      onopen: null,
+      onclose: null,
+      onmessage: null,
+      onerror: null,
+    };
 
-    beforeEach(() => {
-        mockWebSocket = {
-            send: vi.fn(),
-            close: vi.fn(),
-            onopen: null,
-            onclose: null,
-            onmessage: null,
-            onerror: null,
-        };
+    // Mockear el WebSocket global
+    global.WebSocket = vi.fn(() => mockWebSocket);
+  });
 
-        // Mockear el WebSocket global
-        global.WebSocket = vi.fn(() => mockWebSocket);
+  it("Debería abrir el WebSocket y permitir enviar mensajes", async () => {
+    const { result } = renderHook(() => WsHomeService(url));
+
+    await act(async () => {
+      mockWebSocket.onopen();
     });
 
-    it('Debería abrir el WebSocket y permitir enviar mensajes', async () => {
-        const { result } = renderHook(() => WsHomeService(url));
+    expect(result.current.ws).toBe(mockWebSocket);
 
-        await act(async () => {
-            mockWebSocket.onopen(); 
-        });
-
-        expect(result.current.ws).toBe(mockWebSocket);
-        
-        const message = 'Hola Mundo';
-        act(() => {
-            result.current.sendMessage(message);
-        });
-
-        expect(mockWebSocket.send).toHaveBeenCalledWith(message);
+    const message = "Hola Mundo";
+    act(() => {
+      result.current.sendMessage(message);
     });
 
-    it('Debería manejar el cierre del WebSocket', async () => {
-        const { result } = renderHook(() => WsHomeService(url));
+    expect(mockWebSocket.send).toHaveBeenCalledWith(message);
+  });
 
-        await act(async () => {
-            mockWebSocket.onopen();
-        });
+  it("Debería manejar el cierre del WebSocket", async () => {
+    const { result } = renderHook(() => WsHomeService(url));
 
-        await act(async () => {
-            mockWebSocket.onclose();
-        });
-
-        expect(mockWebSocket.close).toHaveBeenCalled();
+    await act(async () => {
+      mockWebSocket.onopen();
     });
 
-    it('No debería enviar mensajes si el WebSocket no está abierto', () => {
-        const { result } = renderHook(() => WsHomeService(url));
-
-        const message = 'Hola Mundo';
-        act(() => {
-            result.current.sendMessage(message);
-        });
-
-        expect(mockWebSocket.send).not.toHaveBeenCalled();
+    await act(async () => {
+      mockWebSocket.onclose();
     });
 
-    it('Debería almacenar mensajes recibidos', async () => {
-        const { result } = renderHook(() => WsHomeService(url));
+    expect(mockWebSocket.close).toHaveBeenCalled();
+  });
 
-        await act(async () => {
-            mockWebSocket.onopen();
-        });
+  it("No debería enviar mensajes si el WebSocket no está abierto", () => {
+    const { result } = renderHook(() => WsHomeService(url));
 
-        const message = 'Mensaje de prueba';
-        await act(async () => {
-            mockWebSocket.onmessage({ data: message });
-        });
-
-        expect(result.current.messages).toContain(message);
+    const message = "Hola Mundo";
+    act(() => {
+      result.current.sendMessage(message);
     });
 
-    afterEach(() => {
-        vi.clearAllMocks();
+    expect(mockWebSocket.send).not.toHaveBeenCalled();
+  });
+
+  it("Debería almacenar mensajes recibidos", async () => {
+    const { result } = renderHook(() => WsHomeService(url));
+
+    await act(async () => {
+      mockWebSocket.onopen();
     });
-    afterAll(() => {
-        consoleLogSpy.mockRestore();
-        consoleErrorSpy.mockRestore();
+
+    const message = "Mensaje de prueba";
+    await act(async () => {
+      mockWebSocket.onmessage({ data: message });
     });
+
+    expect(result.current.messages).toContain(message);
+  });
+
+  afterEach(() => {
+    vi.clearAllMocks();
+  });
+  afterAll(() => {
+    consoleLogSpy.mockRestore();
+    consoleErrorSpy.mockRestore();
+  });
 });
